requirements.txt
Flask==3.0.3
google-api-python-client==2.149.0
python-dotenv==1.0.1
gunicorn==22.0.0
